+++ Basil/include/Scene/Component.h
@@ -95,7 +95,7 @@ namespace Basil
 		BoxCollider2DComponent(const BoxCollider2DComponent& other) = default;
 
 		glm::vec2 offset = { 0.0f, 0.0f };
-		glm::vec2 size = { 0.0f, 0.0f };
+		glm::vec2 size = { 0.5f, 0.5f };
 		float density = 1.0f;
 		float friction = 0.5f;
 		float restitution = 0.0f;